feat: add a new Logo component (and ensure it displays in storybook)
This adds a new logo component, story, and test, while including the
logo on the main Next.JS page.

I noticed that the components weren't displaying in storybook. It was
expecting the .stories files to live inside of `src/components` so I
moved the files in there, which caused them to show up.